+++ app/lib/externalUrlGenerator.js
